refactor: Add src/edusharing-server
