+++ package.json
@@ -29,7 +29,7 @@
     "lint:eslint": "yarn _eslint",
     "lint:prettier": "yarn _prettier --check",
     "lint:tsc": "tsc --noEmit",
-    "start:edusharing": "NODE_OPTIONS='--experimental-specifier-resolution=node' nodemon --watch edusharing-server.ts --watch src/server-utils.ts --watch src/edusharing-server-mock.ts edusharing-server.ts",
+    "start:edusharing": "NODE_OPTIONS='--experimental-specifier-resolution=node' nodemon --watch src/server-utils.ts --watch src/edusharing-server src/edusharing-server/start-server.ts",
     "start:mongodb": "yarn scripts:is-server-running --reverse-mode 27017 && docker-compose up -d || node -e 'process.exit(0)'",
     "start:server": "yarn start:mongodb && TS_NODE_PROJECT='tsconfig.server.json' NODE_OPTIONS='--experimental-specifier-resolution=node' nodemon --watch server.ts --watch src server.ts",
     "start:server-with-edusharing": "npm-run-all -p -l start:edusharing start:server",
